+++ src/index.ts
@@ -1,13 +1,16 @@
 export { createPipe } from './pipe';
 export type { Pipe } from './pipe';
 export type {
+  DefineImpl,
   Fn,
   First,
   Last,
   MergeFns,
   Pipeline,
-  PipeBuilderType,
-  PipeUntyped,
+  PipeCreated,
+  PipeTyped,
+  ResolvedReturnTypes,
   ReturnTypes,
+  TypeSpec,
 } from './types';
 export type { MaybePromiseFn } from './types.return';