+++ src/types.strict.test-d.ts
@@ -1,10 +1,6 @@
 import { createPipe } from './pipe';
-import type {
-  FirstIndexOf,
-  IndexOf,
-  Next,
-  Previous,
-} from './types.strict';
+import type { FirstIndexOf, IndexOf, Previous } from './types.strict';
+import { type as typings } from '@bemedev/typings';
 
 type _Pr = ['a', 'b', 'c', 'd', 'e'];
 expectTypeOf<Previous<_Pr, 0>>().toEqualTypeOf<never>();
@@ -26,24 +22,6 @@ expectTypeOf<Previous<_Pr2, 5>>().toEqualTypeOf<string>();
 expectTypeOf<Previous<_Pr2, 6>>().toEqualTypeOf<never>();
 expectTypeOf<Previous<_Pr2, 7>>().toEqualTypeOf<never>();
 
-type _Nx = ['a', 'b', 'c', 'd', 'e'];
-expectTypeOf<Next<_Nx, 0>>().toEqualTypeOf<'b'>();
-expectTypeOf<Next<_Nx, 1>>().toEqualTypeOf<'c'>();
-expectTypeOf<Next<_Nx, 2>>().toEqualTypeOf<'d'>();
-expectTypeOf<Next<_Nx, 3>>().toEqualTypeOf<'e'>();
-expectTypeOf<Next<_Nx, 4>>().toEqualTypeOf<never>();
-expectTypeOf<Next<_Nx, 5>>().toEqualTypeOf<never>();
-expectTypeOf<Next<_Nx, 6>>().toEqualTypeOf<never>();
-
-type _Nx2 = ['x', 'y', { z: number }, boolean, string, Date];
-expectTypeOf<Next<_Nx2, 0>>().toEqualTypeOf<'y'>();
-expectTypeOf<Next<_Nx2, 1>>().toEqualTypeOf<{ z: number }>();
-expectTypeOf<Next<_Nx2, 2>>().toEqualTypeOf<boolean>();
-expectTypeOf<Next<_Nx2, 3>>().toEqualTypeOf<string>();
-expectTypeOf<Next<_Nx2, 4>>().toEqualTypeOf<Date>();
-expectTypeOf<Next<_Nx2, 5>>().toEqualTypeOf<never>();
-expectTypeOf<Next<_Nx2, 6>>().toEqualTypeOf<never>();
-
 type _IdxOf = ['a', 'b', 'c', 'd', 'e'];
 expectTypeOf<IndexOf<_IdxOf, 'a'>>().toEqualTypeOf<0>();
 expectTypeOf<IndexOf<_IdxOf, 'b'>>().toEqualTypeOf<1>();
@@ -64,11 +42,40 @@ expectTypeOf<FirstIndexOf<_IdxOf2, string>>().toEqualTypeOf<0>();
 expectTypeOf<IndexOf<_IdxOf2, Date>>().toEqualTypeOf<5>();
 expectTypeOf<IndexOf<_IdxOf2, number>>().toEqualTypeOf<never>();
 
-createPipe('add1', 'double', 'add1', 'double', 'add1').init(
+// @ts-expect-error --- no multiple args allowed in first step when keys are duplicated
+createPipe('add1', 'double', 'add1', 'double', 'add1').type<{
+  add1: { parameters: [number, number]; return: number };
+  double: number;
+}>();
+createPipe('add1', 'double', 'add1', 'double', 'add1').type(
   // @ts-expect-error --- no multiple args allowed in first step when keys are duplicated
-  (a: number, b: number) => a + b,
+  typings(({ tuple }) => ({
+    add1: {
+      parameters: tuple('number', 'number'),
+      return: 'number',
+    },
+  })),
 );
 
-createPipe('add1', 'double', 'add1', 'double', 'add1').init(
-  (a: number) => a + 1,
+createPipe('add1', 'double', 'add1', 'double', 'add1').type<{
+  add1: { parameters: [number]; return: number };
+  double: number;
+}>();
+
+createPipe('add1', 'double', 'add1', 'double', 'add1').type<{
+  add1: { parameters: [number]; return: number };
+  double: number;
+}>(
+  typings(({ tuple }) => ({
+    add1: {
+      parameters: tuple('number'),
+      return: 'number',
+    },
+    double: 'number',
+  })),
+);
+
+createPipe('add1', 'double', 'add1', 'double', 'add1').type(
+  // @ts-expect-error ---first key missing
+  typings({}),
 );